feat(logic): mark model
Mark enum model represents the choices players have in the game, either X or O.
This is represented using an enum class as it's a finite set of choices.
Mark is a mixin of both 'str' & 'enum.Enum' & allows the comparison to 'str'
representations of the choices, that is 'X' or 'O'.

The other property is used to represent the only other available choice for Mark.
If X is picked then the only other choice is O. Vice versa holds true.

Additionally, to avoid an `unresolved name` error. annotations import from __future__
package is used to allow using the `Mark` enum class name as the return type on the
`property` method `other`.